backen-michelle/requirements.txt
fastapi>=0.95.0,<0.96.0
pydantic>=1.9.0,<2.0.0
python-dotenv==1.0.1
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
aiomysql>=0.1.0
asyncmy==0.2.8
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
email-validator>=1.1.3,<1.2.0
openai==1.12.0
SpeechRecognition==3.10.0
alembic==1.13.1
bcrypt==4.1.2
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0